requirements.txt
python-telegram-bot==20.8
deep-translator==1.11.4
langdetect==1.0.9
